Fix rangeTo between Float and Long
Float.rangeTo(Long) and Long.rangeTo(Float) now both return DoubleRange. Floats
can store a lot less than longs, so FloatRange is not acceptable. DoubleRange
seems the best we can do here

+++ core/builtins/native/jet/Numbers.kt
@@ -201,7 +201,7 @@ public class Long private () : Number, Comparable<Long> {
   public fun mod(other : Char)   : Long
 
   public fun rangeTo(other : Double) : DoubleRange
-  public fun rangeTo(other : Float)  : FloatRange
+  public fun rangeTo(other : Float)  : DoubleRange
   public fun rangeTo(other : Long)   : LongRange
   public fun rangeTo(other : Int)    : LongRange
   public fun rangeTo(other : Short)  : LongRange
